Fix iOS 8.3 warning

+++ TimesSquare/TSQCalendarCell.m
@@ -97,7 +97,7 @@
     CGFloat extraSpace = (CGRectGetWidth(insetRect) - (self.daysInWeek - 1) * self.columnSpacing) - (increment * self.daysInWeek);
     
     // Divide the extra space out over the outer columns in increments of the column spacing
-    NSInteger columnsWithExtraSpace = (NSInteger)fabsf(extraSpace / self.columnSpacing);
+    NSInteger columnsWithExtraSpace = (NSInteger)fabs(extraSpace / self.columnSpacing);
     NSInteger columnsOnLeftWithExtraSpace = columnsWithExtraSpace / 2;
     NSInteger columnsOnRightWithExtraSpace = columnsWithExtraSpace - columnsOnLeftWithExtraSpace;
